DO NOT USE IN DEVELOPMENT - Implementing use of openssl as PHP suggests
On my linux machine, for some reason the channels aren't decrypting the username and password. As I looked into it, I realised that the way that we are using AES encryption is not very good. So this is mainly looking at how it can be improved while maintaining backwards compatability. This is an experimental feature, don't use it just yet